CallToAction component moved to organisms folder
