+++ src/components/organisms/CallToAction/CallToAction.style.jsx
@@ -1,12 +1,16 @@
 import styled from 'styled-components';
-import ctaBg1 from '../../assets/images/ctaBg1.png';
-import ctaBg2 from '../../assets/images/ctaBg2.png';
-import BREAKPOINTS from '../../shared/constant/breakpoints';
+import ctaBg1 from '../../../assets/images/ctaBg1.png';
+import ctaBg2 from '../../../assets/images/ctaBg2.png';
+import BREAKPOINTS from '../../../shared/constant/breakpoints';
 
 export const StyledCTAWrapper = styled.div`
   display: flex;
   position: relative;
-  height: 500px;
+  height: 400px;
+
+  @media (min-width: ${BREAKPOINTS.md}) {
+    height: 500px;
+  }
 `;
 
 export const StyledCTABgOne = styled.div`
@@ -25,15 +29,14 @@ export const StyledCTABgOne = styled.div`
 
 export const StyledCTABgTwo = styled.div`
   display: none;
+  background-image: url(${ctaBg2});
+  background-repeat: no-repeat;
+  background-position: left center;
+  background-size: cover;
+  width: 50%;
 
   @media (min-width: ${BREAKPOINTS.md}) {
     display: inline-block;
-    width: 50%;
-
-    background-image: url(${ctaBg2});
-    background-repeat: no-repeat;
-    background-position: left center;
-    background-size: cover;
   }
 `;
 
@@ -45,7 +48,7 @@ export const StyledCTAContainer = styled.div`
   padding: 0 60px;
 
   width: 100%;
-  height: 500px;
+  height: 400px;
   background-color: var(--primary);
   align-self: center;
 
@@ -60,9 +63,15 @@ export const StyledCTAContainer = styled.div`
 
   @media (min-width: ${BREAKPOINTS.md}) {
     width: 50%;
+    height: 500px;
   }
 
   @media (min-width: ${BREAKPOINTS.lg}) {
+    margin-left: 300px;
+    width: 600px;
+  }
+
+  @media (min-width: ${BREAKPOINTS.xl}) {
     margin-left: 300px;
     width: 730px;
   }
